requirements.txt

pytest
requests
python-dotenv